requirements.txt
asgiref==3.11.1
gunicorn==22.0.0
brotli==1.2.0
cffi==2.0.0
cssselect2==0.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
et_xmlfile==2.0.0
fonttools==4.62.0
html5lib==1.1
openpyxl==3.1.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==3.0
pydyf==0.12.1
PyJWT==2.11.0
pyphen==0.17.2
python-dateutil==2.9.0
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tzdata==2025.3
weasyprint==62.3
webencodings==0.5.1
zopfli==0.4.1
